+++ OmetriaReactNativeSdk.h
@@ -1,0 +1,12 @@
+//
+//  OmetriaReactNativeSdk.h
+//  Pods
+//
+//  Created by Vlad on 07.02.2024.
+//
+
+#ifndef OmetriaReactNativeSdk_h
+#define OmetriaReactNativeSdk_h
+
+
+#endif /* OmetriaReactNativeSdk_h */