+++ example/metro.config.js
@@ -1,4 +1,5 @@
 const path = require('path');
+const { getDefaultConfig, mergeConfig } = require('@react-native/metro-config');
 const blacklist = require('metro-config/src/defaults/exclusionList');
 const escape = require('escape-string-regexp');
 const pak = require('../package.json');
@@ -9,7 +10,14 @@ const modules = Object.keys({
   ...pak.peerDependencies,
 });
 
-module.exports = {
+/**
+ * Metro configuration
+ * https://facebook.github.io/metro/docs/configuration
+ *
+ * @type {import('metro-config').MetroConfig}
+ */
+
+const config = {
   projectRoot: __dirname,
   watchFolders: [root],
 
@@ -38,3 +46,5 @@ module.exports = {
     }),
   },
 };
+
+module.exports = mergeConfig(getDefaultConfig(__dirname), config);