+++ ios/OmetriaReactNativeSdk.m
@@ -9,6 +9,11 @@ RCT_EXTERN_METHOD(initializeWithApiToken:(NSString *)apiToken
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)
 
+RCT_EXTERN_METHOD(initializeWithApiToken:(NSString *)apiToken
+                  options:(NSDictionary *)options
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+
 RCT_EXTERN_METHOD(trackProfileIdentifiedByCustomerIdEvent:(NSString *)customerId
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)
@@ -29,10 +34,11 @@ RCT_EXTERN_METHOD(trackProductListingViewedEvent:(NSString *)listingType
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)
 
+// Deprecated
 RCT_EXTERN_METHOD(trackWishlistAddedToEvent:(NSString *)productId
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)
-
+//Deprecated
 RCT_EXTERN_METHOD(trackWishlistRemovedFromEvent:(NSString *)productId
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)
@@ -41,6 +47,14 @@ RCT_EXTERN_METHOD(trackBasketViewedEvent:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)
 
 RCT_EXTERN_METHOD(trackBasketUpdatedEvent:(NSDictionary *)basket
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+
+RCT_EXTERN_METHOD(onNotificationInteracted:(NSDictionary *)remoteMessage
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+
+RCT_EXTERN_METHOD(onNotificationReceived:(NSDictionary *)remoteMessage
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)